backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
pyyaml==6.0.1
celery==5.3.4
kombu==5.3.4
prometheus-client==0.19.0
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2
itsdangerous==2.1.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pillow==10.1.0